+++ CHANGELOG.md
@@ -50,6 +50,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implement stubs for `LaunchListResponse'`, `LaunchListItem`, `LaunchResponse`, `LaunchServiceProviderResponse`, `PadResponse`, `LocationResponse`.
 - `LaunchListDataSource` and `LaunchListTableViewCell`.
 - `default_rocket_launch_image.jpg` to the Assets.
+- custom `LoadingTableFooterView` for the prefetching indication.
 
 ### Changed